Move/renamespace DecoratedSuccessResponse
